+++ .gitignore
@@ -9,7 +9,8 @@
 !.github/workflows/utest.yml
 
 !srcs
-!srcs/*.py
+!srcs/modules
+!srcs/**/**.py
 
 !tests
 !tests/*.py
