+++ tests/test_sequential.py
@@ -3,13 +3,13 @@ import sys
 sys.path.append(os.pardir)
 
 import numpy as np
-from srcs.functions import Softmax
-from srcs.activation import ReLU, Sigmoid
-from srcs.loss import CrossEntropyLoss
-from srcs.init import he_normal, xavier_normal, normal
-from srcs.optimizer import SGD, Momentum, Nesterov, AdaGrad, RMSProp, Adam
-from srcs.layer import Dense
-from srcs.model import Sequential
+from srcs.modules.functions import Softmax
+from srcs.modules.activation import ReLU, Sigmoid
+from srcs.modules.loss import CrossEntropyLoss
+from srcs.modules.init import he_normal, xavier_normal, normal
+from srcs.modules.optimizer import SGD, Momentum, Nesterov, AdaGrad, RMSProp, Adam
+from srcs.modules.layer import Dense
+from srcs.modules.model import Sequential
 
 from sklearn.model_selection import train_test_split
 from keras.datasets import mnist
